+++ lib/database/entities/country.dart
@@ -3,7 +3,7 @@ import 'package:floor/floor.dart';
 @Entity(tableName: 'country')
 class Country {
   @PrimaryKey(autoGenerate: true)
-  final int id;
+  final int? id;
 
   @ColumnInfo(name: 'name')
   final String name;