+++ lib/game_lists.dart
@@ -2,14 +2,19 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:game_lists/database/database.dart';
 import 'package:game_lists/pages/about_page.dart';
-import 'package:game_lists/pages/country_page.dart';
-import 'package:game_lists/pages/developer_page.dart';
-import 'package:game_lists/pages/game_page.dart';
+import 'package:game_lists/pages/add_edit_pages/country_page.dart';
+import 'package:game_lists/pages/add_edit_pages/developer_page.dart';
+import 'package:game_lists/pages/select_pages/select_country_page.dart';
+import 'package:game_lists/pages/select_pages/select_developer_page.dart';
+import 'package:game_lists/pages/add_edit_pages/game_page.dart';
 import 'package:game_lists/pages/main_page.dart';
-import 'package:game_lists/pages/platform_page.dart';
-import 'package:game_lists/pages/series_page.dart';
+import 'package:game_lists/pages/add_edit_pages/platform_page.dart';
+import 'package:game_lists/pages/add_edit_pages/series_page.dart';
+import 'package:game_lists/pages/select_pages/select_platform_page.dart';
+import 'package:game_lists/pages/select_pages/select_series_page.dart';
+import 'package:game_lists/pages/select_pages/select_tag_page.dart';
 import 'package:game_lists/pages/settings_page.dart';
-import 'package:game_lists/pages/tag_page.dart';
+import 'package:game_lists/pages/add_edit_pages/tag_page.dart';
 
 class GameLists extends StatefulWidget {
   const GameLists({Key? key, this.database}) : super(key: key);
@@ -31,15 +36,23 @@ class _GameListsState extends State<GameLists> {
       ),
       initialRoute: '/',
       routes: {
+        // Main page
         '/': (context) => MainPage(database: widget.database),
         '/about': (context) => AboutPage(),
+        '/settings': (context) => SettingsPage(),
+        // Add Edit Page
         '/country': (context) => CountryPage(),
         '/developer': (context) => DeveloperPage(),
         '/game': (context) => GamePage(),
         '/platform': (context) => PlatformPage(),
         '/series': (context) => SeriesPage(),
-        '/settings': (context) => SettingsPage(),
         '/tag': (context) => TagPage(),
+        // Select Pages
+        '/select_country': (context) => SelectCountryPage(),
+        '/select_developer': (context) => SelectDeveloperPage(),
+        '/select_platform': (context) => SelectPlatformPage(),
+        '/select_series': (context) => SelectSeriesPage(),
+        '/select_tag': (context) => SelectTagPage(),
       },
       debugShowCheckedModeBanner: false,
     );