+++ lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:game_lists/database/database.dart';
 import 'package:game_lists/game_lists.dart';
+import 'package:provider/provider.dart';
 
 void main() async {
   SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
@@ -13,5 +14,8 @@ void main() async {
   final database = await $FloorGameListsDatabase
       .databaseBuilder('game_lists_database.db')
       .build();
-  runApp(GameLists(database: database));
+  runApp(Provider<GameListsDatabase>(
+    create: (_) => database,
+    child: GameLists(database: database),
+  ));
 }